enforce one instance of Rclone Browser per user
The lock file is created in a temporary directory that is specific to a logged in user. This means that each user on the system will be able to run one instance of the app for it’s own.

There is no danger when the app crashes, because the lock will be destroyed with the app instance.