set app locale to UK english - #96
I hoped it is avoidable and Qt can handle locale but it leads to issues like #96.  If one day somebody undertakes RB localisation it can be solved in nicer way. But from now on chaps - english only. App does not use any input/output date/currency etc. so should be not noticeable (it is in english only anyway)